feat: need to revisit 2093B, very interesting solution

+++ template.cpp
@@ -11,7 +11,7 @@
 #define S second
 #define PB push_back
 #define MP make_pair
-#define REP(i, a, b) for (size_t i = a; i < b; i++)
+#define REP(i, a, b) for (int i = a; i < b; i++)
 
 using namespace std;
